more protocol re-factor
