+++ docs/spec/data_message.spec
