--- protocols/dna_assembly_raf/run.sh
@@ -1,5 +1,0 @@
-./lambda6_listen.sh
-./lambda6_listen_5556.sh
-./lambda6_listen_5557.sh
-./lambda6_listen_5558.sh
-./lambda6_listen_5559.sh
