+++ protocols/dna_assembly_raf/run_protocol.py
@@ -1,28 +1,15 @@
+import argparse
+
+from .post_transformation import generate_post_transformation
 
 
-
-from . import step1
-from . import step2
-
-
-full_protocol = [
-    step1,
-    step2,
-    ]
-
-
-
+def parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-t", "--is_test", help="use -t or --is_test only if the run is a test and the data can be deleted", action="store_true")
+    return parser.parse_args()
 
 if __name__ == "__main__":
-    # handle command line args
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "-t", 
-        "--is_test",
-        help="use -t or --is_test only if the run is a test and the data can be deleted",  
-        action="store_true",
-    )
-    args = vars(parser.parse_args())
 
-    # pass to method
+    args = parse_args()
+
     generate_post_transformation(args["is_test"])
