+++ protocols/dna_assembly_raf/utils/send_protocol.py
@@ -1,0 +1,20 @@
+from subprocess import Popen
+
+def send_protocol(dir_path, i_list):
+    try:
+        # TODO: change to full path on lambda6
+        child_message_sender = child_pid = Popen(
+            [
+                "python",
+                "../../zeromq/lambda6_send_instructions.py",
+                "-d",
+                dir_path,
+                "-i"
+                ].extend(i_list),
+            start_new_session=True,
+        ).pid
+
+        print("New instruction directory passed to lambda6_send_message.py")
+    except BaseException as e:
+        print(e)
+        print("Could not send new instructions to hudson01")
