+++ protocols/dna_assembly_raf/utils/wc_setup.py
@@ -1,4 +1,5 @@
-
+from liquidhandling import SoftLinx
+from .tip_utils import replace_tip_box
 
 # HELPER METHODS: ---------------------------------------------------------------
 def set_up(
