+++ rdbms/db_config.py
@@ -1,0 +1,5 @@
+# Database variables
+DBNAME = "name"
+DBUSER = "user"
+DBPASSWD = "password"
+DBHOST = "host"
